rename EntityRenderDispatcherExtension to VR part 2
