+++ common/src/main/java/org/vivecraft/mixin/client/renderer/entity/EntityRenderDispatcherVRMixin.java
@@ -1,7 +1,7 @@
 package org.vivecraft.mixin.client.renderer.entity;
 
 import org.vivecraft.ClientDataHolder;
-import org.vivecraft.extensions.EntityRenderDispatcherExtension;
+import org.vivecraft.extensions.EntityRenderDispatcherVRExtension;
 import org.vivecraft.extensions.LevelRendererExtension;
 import com.mojang.math.Quaternion;
 import com.mojang.math.Vector3f;
@@ -28,7 +28,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 @Mixin(EntityRenderDispatcher.class)
-public abstract class EntityRenderDispatcherVRMixin implements ResourceManagerReloadListener, EntityRenderDispatcherExtension {
+public abstract class EntityRenderDispatcherVRMixin implements ResourceManagerReloadListener, EntityRenderDispatcherVRExtension {
 
     @Unique
     private VRArmRenderer armRenderer;
